--- events/src/testFixtures/java/tech/pegasys/teku/events/TestExceptionHandler.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2020 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.events;
-
-import java.lang.reflect.Method;
-import org.assertj.core.api.Assertions;
-
-public enum TestExceptionHandler implements ChannelExceptionHandler {
-  TEST_EXCEPTION_HANDLER;
-
-  @Override
-  public void handleException(
-      final Throwable error,
-      final Object subscriber,
-      final Method invokedMethod,
-      final Object[] args) {
-    Assertions.fail("Unhandled error in subscriber ...", error);
-  }
-}
