+++ services/chainstorage/build.gradle
@@ -7,7 +7,7 @@ dependencies {
   implementation project(':storage:api')
   implementation project(':util')
   implementation project(':pow')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':protoarray')
 
   implementation 'org.apache.tuweni:tuweni-config'
