+++ services/remote-validator/build.gradle
@@ -4,7 +4,7 @@ dependencies {
   implementation project(':data:serializer')
   implementation project(':ethereum:statetransition')
   implementation project(':ethereum:datastructures')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':services:serviceutils')
   implementation project(':storage:api')
