+++ services/serviceutils/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
   implementation project(':data:metrics')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:logging')
   implementation project(':util')
