+++ services/timer/build.gradle
@@ -2,7 +2,7 @@ dependencies {
     implementation project(':services:serviceutils')
     implementation project(':infrastructure:async')
     implementation project(':util')
-    implementation project(':events')
+    implementation project(':infrastructure:events')
     implementation 'com.google.guava:guava'
     implementation('org.quartz-scheduler:quartz') {
         exclude group: 'com.mchange'
