+++ settings.gradle
@@ -13,10 +13,10 @@ include 'data:recorder'
 include 'data:serializer'
 include 'data:signingrecord'
 include 'data:yaml'
-include 'events'
 include 'fuzz'
 include 'infrastructure'
 include 'infrastructure:async'
+include 'infrastructure:events'
 include 'infrastructure:io'
 include 'infrastructure:unsigned'
 include 'infrastructure:logging'
