+++ sync/build.gradle
@@ -13,7 +13,7 @@ dependencies {
     implementation project(':storage')
     implementation project(':storage:api')
     implementation project(':util')
-    implementation project(':events')
+    implementation project(':infrastructure:events')
 
     implementation 'org.apache.tuweni:tuweni-bytes'
 
@@ -41,7 +41,7 @@ dependencies {
     testFixturesImplementation testFixtures(project(':infrastructure:unsigned'))
     testFixturesImplementation testFixtures(project('::networking:eth2'))
     testFixturesImplementation testFixtures(project('::networking:p2p'))
-    testFixturesImplementation testFixtures(project('::events'))
+    testFixturesImplementation testFixtures(project('::infrastructure:events'))
 
     testFixturesImplementation 'org.hyperledger.besu:plugin-api'
     testFixturesImplementation 'org.hyperledger.besu.internal:metrics-core'
