+++ sync/src/testFixtures/java/tech/pegasys/teku/sync/SyncingNodeManager.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.teku.sync;
 
-import static tech.pegasys.teku.events.TestExceptionHandler.TEST_EXCEPTION_HANDLER;
+import static tech.pegasys.teku.infrastructure.events.TestExceptionHandler.TEST_EXCEPTION_HANDLER;
 
 import com.google.common.eventbus.EventBus;
 import java.util.List;
@@ -22,9 +22,9 @@ import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.core.StateTransition;
 import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.events.EventChannels;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.events.EventChannels;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.eth2.Eth2Network;
 import tech.pegasys.teku.networking.eth2.Eth2NetworkFactory;
