+++ teku/build.gradle
@@ -7,7 +7,7 @@ dependencies {
   implementation project(':ethereum:core')
   implementation project(':ethereum:datastructures')
   implementation project(':ethereum:statetransition')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:logging')
   implementation project(':networking:p2p')
