+++ validator/client/build.gradle
@@ -5,7 +5,7 @@ dependencies {
   implementation project(':data:metrics')
   implementation project(':ethereum:core')
   implementation project(':ethereum:datastructures')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:logging')
   implementation project(':services:serviceutils')
