+++ validator/eventadapter/build.gradle
@@ -3,7 +3,7 @@ dependencies {
   implementation project(':bls')
   implementation project(':ethereum:datastructures')
   implementation project(':ethereum:statetransition')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':services:serviceutils')
   implementation project(':storage:api')
