+++ validator/remote/build.gradle
@@ -2,7 +2,7 @@ dependencies {
   implementation project(':bls')
   implementation project(':data:serializer')
   implementation project(':ethereum:datastructures')
-  implementation project(':events')
+  implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':services:serviceutils')
   implementation project(':storage:api')
